[Feat] CUSER_008 완료, CUSER_004 완료, CUSER_005 진행

+++ src/main/java/com/fiiiiive/zippop/member/MemberService.java
@@ -5,16 +5,28 @@ import com.fiiiiive.zippop.common.responses.BaseResponseMessage;
 import com.fiiiiive.zippop.member.model.Company;
 import com.fiiiiive.zippop.member.model.CustomUserDetails;
 import com.fiiiiive.zippop.member.model.Customer;
+import com.fiiiiive.zippop.member.model.request.EditInfoReq;
+import com.fiiiiive.zippop.member.model.request.EditPasswordReq;
 import com.fiiiiive.zippop.member.model.request.PostSignupReq;
 import com.fiiiiive.zippop.member.model.response.PostSignupRes;
+import com.google.gson.Gson;
 import lombok.RequiredArgsConstructor;
-import org.springframework.mail.MailException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.ResponseEntity;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
+import org.springframework.web.client.RestTemplate;
 
+import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
@@ -23,6 +35,7 @@ import java.util.UUID;
 @Service
 @RequiredArgsConstructor
 public class MemberService {
+    private static final Logger log = LoggerFactory.getLogger(MemberService.class);
     private final JavaMailSender emailSender;
     private final CompanyRepository companyRepository;
     private final CustomerRepository customerRepository;
@@ -71,7 +84,7 @@ public class MemberService {
             Optional<Customer> result = customerRepository.findByEmail(request.getEmail());
             if(result.isPresent()){
                 Customer customer = result.get();
-                if(customer.getEnabled()) {
+                if(customer.getEnabled() && customer.getPassword() == null) {
                     throw new BaseException(BaseResponseMessage.MEMBER_REGISTER_FAIL_ALREADY_EXIST);
                 } else {
                     return PostSignupRes.builder()
@@ -101,7 +114,7 @@ public class MemberService {
         }
     }
 
-    public Boolean activeMember(String email, String role) throws Exception, BaseException {
+    public Boolean activeMember(String email, String role) throws BaseException {
         if(Objects.equals(role, "ROLE_COMPANY")){
             Optional<Company> result = companyRepository.findByEmail(email);
             if(result.isPresent()){
@@ -124,7 +137,7 @@ public class MemberService {
         return true;
     }
 
-    public String sendEmail(PostSignupRes response) throws MailException, Exception {
+    public String sendEmail(PostSignupRes response) {
         SimpleMailMessage message = new SimpleMailMessage();
         message.setTo(response.getEmail());
         if(Objects.equals(response.getRole(), "ROLE_COMPANY")){
@@ -149,8 +162,9 @@ public class MemberService {
 
     @Transactional(rollbackFor = Exception.class)
     public void inActiveMenber(CustomUserDetails customUserDetails) throws BaseException {
-        String email = customUserDetails.getEmail();
+        String email = customUserDetails.getUsername();
         String role = customUserDetails.getRole();
+        Long idx = customUserDetails.getIdx();
         if(Objects.equals(role, "ROLE_COMPANY")){
             Optional<Company> result = companyRepository.findByEmail(email);
             if(result.isPresent()){
@@ -166,11 +180,83 @@ public class MemberService {
             if(result.isPresent()) {
                 Customer customer = result.get();
                 customer.setEnabled(false);
-                customerRepository.save(customer);
-                emailVerifyRepository.deleteByEmail(email);
+                if(customer.getPassword() == null){
+                    customerRepository.save(customer);
+                } else {
+                    customerRepository.save(customer);
+                    emailVerifyRepository.deleteByEmail(email);
+                }
             } else {
                 throw new BaseException(BaseResponseMessage.MEMBER_INACTIVE_FAIL);
             }
         }
     }
+
+    @Transactional(rollbackFor = Exception.class)
+    public void editInfo(CustomUserDetails customUserDetails, EditInfoReq dto) throws BaseException {
+        String email = customUserDetails.getEmail();
+        String role = customUserDetails.getRole();
+        if(Objects.equals(role, "ROLE_COMPANY")){
+            Optional<Company> result = companyRepository.findByEmail(email);
+            if(result.isPresent()){
+                Company company = result.get();
+                company.setName(dto.getName());
+                company.setAddress(dto.getAddress());
+                company.setCrn(dto.getCrn());
+                company.setPhoneNumber(dto.getPhoneNumber());
+                companyRepository.save(company);
+            } else {
+                throw new BaseException(BaseResponseMessage.MEMBER_EDIT_INFO_FAIL);
+            }
+        } else {
+            Optional<Customer> result = customerRepository.findByEmail(customUserDetails.getEmail());
+            if(result.isPresent()) {
+                Customer customer = result.get();
+                customer.setName(dto.getName());
+                customer.setAddress(dto.getAddress());
+                customer.setPhoneNumber(dto.getPhoneNumber());
+                customerRepository.save(customer);
+            } else {
+                throw new BaseException(BaseResponseMessage.MEMBER_EDIT_INFO_FAIL);
+            }
+        }
+    }
+
+    @Transactional(rollbackFor = Exception.class)
+    public void editPassword(CustomUserDetails customUserDetails, EditPasswordReq dto) throws BaseException {
+        String email = customUserDetails.getEmail();
+        String role = customUserDetails.getRole();
+        if(Objects.equals(role, "ROLE_COMPANY")){
+            Optional<Company> result = companyRepository.findByEmail(email);
+            if(result.isPresent()){
+                Company company = result.get();
+                if(passwordEncoder.matches(dto.getOriginPassword(), company.getPassword()))
+                {
+                    company.setPassword(passwordEncoder.encode(dto.getNewPassword()));
+                    companyRepository.save(company);
+                }
+                else {
+                    throw new BaseException(BaseResponseMessage.MEMBER_EDIT_PASSWORD_FAIL_PASSWORD_NOT_MATCH);
+                }
+            } else {
+                throw new BaseException(BaseResponseMessage.MEMBER_EDIT_PASSWORD_FAIL);
+            }
+        } else {
+            Optional<Customer> result = customerRepository.findByEmail(customUserDetails.getEmail());
+            if(result.isPresent()) {
+                Customer customer = result.get();
+                if(passwordEncoder.matches(dto.getOriginPassword(), customer.getPassword()))
+                {
+                    customer.setPassword(passwordEncoder.encode(dto.getNewPassword()));
+                    customerRepository.save(customer);
+                }
+                else {
+                    throw new BaseException(BaseResponseMessage.MEMBER_EDIT_PASSWORD_FAIL_PASSWORD_NOT_MATCH);
+                }
+            } else {
+                throw new BaseException(BaseResponseMessage.MEMBER_EDIT_PASSWORD_FAIL);
+            }
+        }
+    }
+
 }
